Move MeshParameter and model_h from Mesh/ to Utils/
MeshParameter.py and model_h.skops implement the stabilization scheme
used by mixed.py, not mesh/grid logic. They belong in Utils as shared
ingredients, not in Mesh alongside Grid.py.

Changes:
- Move files: Mesh/MeshParameter.py, Mesh/model_h.skops → Utils/
- Update internal import in MeshParameter (from ..Utils.IO → from .IO)
- Update Mesh/__init__.py to remove MeshParameter exports
- Update Utils/__init__.py to export Element, ModelML
- Update mixed.py import to use Utils.MeshParameter
- Update pyproject.toml package-data path

Verified: imports work from new locations, model_h.skops loads correctly
