+++ safeincave/Mesh/__init__.py
@@ -3,6 +3,5 @@
 # SPDX-License-Identifier: BSD-3-Clause
 
 from .Grid import GridHandlerGMSH
-from .MeshParameter import Element, ModelML
 
-__all__ = ["GridHandlerGMSH", "Element", "ModelML"]
+__all__ = ["GridHandlerGMSH"]
